requirements.txt
asyncpg==0.19.0
Django==2.2.6
psycopg2-binary==2.8.3
pytz==2019.3
PyYAML==5.1.2
sqlparse==0.3.0
